requirements.txt
# Core Dependencies
torch>=2.1.0
git+https://github.com/huggingface/transformers.git
accelerate>=0.25.0
bitsandbytes>=0.41.0

# LangChain Ecosystem
langchain>=0.1.0
langchain-community>=0.0.20
langchain-huggingface>=0.0.1
langchain-text-splitters>=0.0.1
langchain-chroma>=0.1.0

# Vector Store & Embeddings
chromadb>=0.4.22
sentence-transformers>=2.3.0

# Document Processing
pypdf>=4.0.0
PyMuPDF>=1.23.0  # Better PDF handling with image extraction
pdf2image>=1.16.3  # Convert PDF pages to images
pillow>=10.0.0  # Image processing

# OCR & Table Detection
paddleocr>=2.7.0  # Main OCR engine
paddlepaddle>=2.5.0  # PaddlePaddle framework (CPU version)
# paddlepaddle-gpu>=2.5.0  # Use this if you have CUDA (uncomment and remove CPU version)

# Table Processing
pdfplumber>=0.10.0  # Better table extraction
tabulate>=0.9.0  # Format tables nicely

# UI
gradio>=4.0.0

# Utilities
protobuf==3.20.3
scipy>=1.11.0
tiktoken>=0.5.0
einops>=0.7.0
numpy>=1.24.0
mistral-common>=1.5.0
sentencepiece>=0.1.99
sacremoses>=0.0.53
opencv-python>=4.8.0  # Required by PaddleOCR
shapely>=2.0.0  # Geometry operations for OCR
